Introduce and `RSAPublicKey` type.
For now used internally in place of passing around an `(n, e)` pair, and in
place of the `n` and `e` fields in `RSAKeyPair`.

Some of the `src/rsa/rsa.rs` functions and `RSAKeyPair` methods pertaining to
the public key components have been moved to `RSAPublicKey` methods.

Parts of `check_public_modulus_and_exponent` are now found in the
`RSAPublicKey` constructors, while the `n_min_bits` and `n_max_bits` checks are
still in a separate `check_modulus` method. (The parameters differ between
verification and signing.)