Eliminate `VerificationAlgorithm` trait.
The original intent here was to add a `verify` variant that accepted one of the
`*PublicKey` structs, which would entail adding an associated type. This seems
to complicate things a lot, however, especially if the type has a lifetime
parameter.

Instead, eliminate `VerificationAlgorithm` altogether, and bring the
verification API closer to what's already there for signing.

This also eliminates the unit struct `EdDSAParameters`.